climb detail cleanup
